ckb-auth-cli, command related code is moved to chain_command
