+++ tools/ckb-auth-cli/src/chain_command/bitcoin.rs
@@ -1,5 +1,4 @@
-use super::{BlockChain, BlockChainArgs};
-use crate::utils::decode_string;
+use crate::{auth_script::run_auth_exec, utils::decode_string, BlockChain, BlockChainArgs};
 use anyhow::{anyhow, Error};
 use ckb_auth_types::AuthAlgorithmIdType;
 use clap::{arg, ArgMatches, Command};
@@ -75,7 +74,7 @@ impl BlockChain for BitcoinLock {
 
         let pubkey_hash = &address[1..21];
 
-        super::auth_script::run_auth_exec(
+        run_auth_exec(
             AuthAlgorithmIdType::Bitcoin,
             pubkey_hash,
             &message,
